+++ src/reducers/reducer/reducer.test.js
@@ -1,4 +1,4 @@
-import {reducer} from './reducer';
+import reducer from './reducer';
 
 describe(`Reducer works correctly`, () => {
